requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.0
ffmpeg-python==0.2.0
requests==2.31.0 